Added option to run redundant tests (#279)
* Adds option to run redundant tests.

Sometimes tests are incorrectly deemed redundant.
In those cases a cheap fix is to simply run all tests
for every candidate.

* Reverts image name.

A confused person (me) didn't realize there
were example files within Darjeeling and was
using only the example files in a separate
repo. This lead the lead author to fix what the
confused person (me) stated was a bug.

* Fixes incompatible argument type.

Needs to be bool, not Optional[bool].

* Fixes formatting mistakes.

Trailing spaces and incorrectly indented code.

* Changes '_' to '-' in YAML for consistency.

`run_redundant_tests' is now 'run-redundant-tests' for
consistency with the other options.